Extract menuOnContextMenu to views helper
