+++ src/ui/views/Table/components/DataGrid/dataGrid.ts
@@ -1,4 +1,3 @@
-import type { Menu } from "obsidian";
 import type { DataFieldType, DataField } from "src/lib/dataframe/dataframe";
 
 export type GridValidRowModel = { [key: string]: any };
@@ -18,14 +17,4 @@ export type GridRowId = string;
 export interface GridRowProps {
   readonly rowId: GridRowId;
   readonly row: GridRowModel;
-}
-
-export function menuOnContextMenu(event: MouseEvent, menu: Menu): void {
-  const contextMenuFunc = (event: MouseEvent) => {
-    window.removeEventListener("contextmenu", contextMenuFunc);
-    event.preventDefault();
-    event.stopPropagation();
-    menu.showAtMouseEvent(event);
-  };
-  window.addEventListener("contextmenu", contextMenuFunc, false);
 }
