+++ src/ui/views/helpers.ts
@@ -1,6 +1,7 @@
 import { makeContext } from "src/lib/helpers";
 import { DataFieldType, type DataField } from "../../lib/dataframe/dataframe";
 import type { ViewProps } from "../app/useView";
+import type { Menu } from "obsidian";
 
 export function fieldIcon(field: DataFieldType): string {
   switch (field) {
@@ -28,3 +29,13 @@ export function fieldToSelectableValue(field: DataField): {
 
 export const getRecordColorContext = makeContext<ViewProps["getRecordColor"]>();
 export const sortRecordsContext = makeContext<ViewProps["sortRecords"]>();
+
+export function menuOnContextMenu(event: MouseEvent, menu: Menu): void {
+  const contextMenuFunc = (event: MouseEvent) => {
+    window.removeEventListener("contextmenu", contextMenuFunc);
+    event.preventDefault();
+    event.stopPropagation();
+    menu.showAtMouseEvent(event);
+  };
+  window.addEventListener("contextmenu", contextMenuFunc, false);
+}
